Removed Corbett as at-risk, adjusted course colors accordingly
Means that the graduated red colors had to change for a bunch of
classes. Many reverted to their parent's color, others just decremented
by one.

Realized (but haven't fixed) that the annotation box placement is
absolute, when it should be relative to the center of the browser
window. Don't yet know how to do that.