Playing with position of annotation
Made it relative to the right edge of the browser instead of the left.
And reduced the width to reduce risk of overlap with the wheel. When the
string is too long, it automatically makes a taller box for me.

Would prefer to make the annotation position relative to the center of
the wheel, but don't know how to do that yet.